Updated player animation triggers, wall climb, and player states
Renamed and updated player animation triggers, and also edited the wall climb animation. Player states were updated to add a check for if the player is touching a ledge, and variables were renamed for better clarity. Other changes include the modification of player touchdown state, idle state, and a host of others to provide better user experience. Additionally, a new image for ledge grabbing was added.